Update Mycsv.php
The original code from fetch_assoc function is moved to __construct to give an instant num_rows and validation of the query. Also, now you can SELECT the columns in random order and it will give you the selected columns back in order of the table. 

Meanwhile, there is a dead mosquito on my hand